+++ expui/src/main/kotlin/io/kanro/compose/jetbrains/expui/style/TextStyle.kt
@@ -1,9 +1,52 @@
 package io.kanro.compose.jetbrains.expui.style
 
 import androidx.compose.runtime.compositionLocalOf
-import androidx.compose.ui.text.TextStyle
 import io.kanro.compose.jetbrains.expui.theme.LightTheme
 
-val LocalTextStyle = compositionLocalOf<TextStyle> {
+val LocalDefaultTextStyle = compositionLocalOf {
     LightTheme.DefaultTextStyle
+}
+
+val LocalDefaultBoldTextStyle = compositionLocalOf {
+    LightTheme.DefaultBoldTextStyle
+}
+
+val LocalParagraphTextStyle = compositionLocalOf {
+    LightTheme.ParagraphTextStyle
+}
+
+val LocalMediumTextStyle = compositionLocalOf {
+    LightTheme.MediumTextStyle
+}
+
+val LocalMediumBoldTextStyle = compositionLocalOf {
+    LightTheme.MediumBoldTextStyle
+}
+
+val LocalSmallTextStyle = compositionLocalOf {
+    LightTheme.SmallTextStyle
+}
+
+val LocalH0TextStyle = compositionLocalOf {
+    LightTheme.H0TextStyle
+}
+
+val LocalH1TextStyle = compositionLocalOf {
+    LightTheme.H1TextStyle
+}
+
+val LocalH2TextStyle = compositionLocalOf {
+    LightTheme.H2TextStyle
+}
+
+val LocalH2BoldTextStyle = compositionLocalOf {
+    LightTheme.H2BoldTextStyle
+}
+
+val LocalH3TextStyle = compositionLocalOf {
+    LightTheme.H3TextStyle
+}
+
+val LocalH3BoldTextStyle = compositionLocalOf {
+    LightTheme.H3BoldTextStyle
 }