+++ expui/src/main/kotlin/io/kanro/compose/jetbrains/expui/theme/Theme.kt
@@ -1,6 +1,8 @@
 package io.kanro.compose.jetbrains.expui.theme
 
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.CompositionLocalProvider
+import androidx.compose.runtime.ProvidedValue
 import androidx.compose.runtime.compositionLocalOf
 
 val LocalIsDarkTheme = compositionLocalOf { false }
@@ -9,5 +11,12 @@ interface Theme {
     val isDark: Boolean
 
     @Composable
-    fun provide(content: @Composable () -> Unit)
+    fun provide(content: @Composable () -> Unit) {
+        CompositionLocalProvider(
+            *provideValues(),
+            content = content,
+        )
+    }
+
+    fun provideValues(): Array<ProvidedValue<*>>
 }