+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
 ## [Unreleased]
+- Provide BasicMainToolBar #18
+- Disable minimize and maximize button when resizeable is false #19
+- Fix the custom title bar when MainToolBar Content is empty #20
+- Optimize the implementation of themes. Switching themes will not cause full recompose.
+- Add text styles for theme
 
 ## [2.0.0]
 ## What's Changed